pyproject.toml
[project]
name = "syukatsu-support"
version = "0.1.0"
description = "Syukatsu Support Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai==2.15.0",
    "pydantic>=2.0.0",
    "pydantic-settings",
    "python-dotenv",
    "cryptography",
    "tenacity",
    "structlog>=25.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]
